Added basic card info
