+++ checkoutcalculator/app/src/App.js
@@ -3,7 +3,6 @@ import logo from './logo.svg';
 import './App.css';
 import Navigation from './components/Navbar';
 import Items from './components/Items';
-import Cart from './components/Cart';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 function App() {
@@ -11,7 +10,6 @@ function App() {
     <div>
       <Navigation />
       <Items />
-      <Cart />
     </div>
   );
 }
